refactor: Split SectionsLayout into topic files
