+++ app/src/main/java/app/morphe/manager/ui/screen/home/SectionsLayout.kt
@@ -9,7 +9,6 @@ import android.annotation.SuppressLint
 import android.content.pm.PackageInfo
 import android.view.HapticFeedbackConstants
 import androidx.activity.compose.BackHandler
-import androidx.appcompat.content.res.AppCompatResources
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.Crossfade
@@ -46,7 +45,6 @@ import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.hapticfeedback.HapticFeedbackType
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.layout.boundsInWindow
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.*
@@ -65,17 +63,12 @@ import app.morphe.manager.domain.manager.HomeAppCategory
 import app.morphe.manager.domain.manager.HomeAppCategoryState
 import app.morphe.manager.domain.manager.HomeAppCategoryViewMode
 import app.morphe.manager.domain.manager.HomeAppSortMode
-import app.morphe.manager.domain.repository.PatchBundleRepository
 import app.morphe.manager.patcher.patch.PatchInfo
 import app.morphe.manager.ui.model.HomeAppItem
 import app.morphe.manager.ui.screen.shared.*
-import app.morphe.manager.ui.viewmodel.BundleUpdateStatus
 import app.morphe.manager.ui.viewmodel.HomeAppSourceGroup
 import app.morphe.manager.util.AppDataSource
 import app.morphe.manager.util.KnownApps
-import app.morphe.manager.util.RemoteAvatar
-import app.morphe.manager.util.toast
-import com.google.accompanist.drawablepainter.rememberDrawablePainter
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import sh.calvin.reorderable.ReorderableItem
@@ -90,51 +83,8 @@ private data class SwipeActionConfig(
     val contentColor: Color
 )
 
-/**
- * UI-side group backing one collapsible section in the home list. Represents either a
- * user-defined category (has [id], [editable] true, no source fields), a source group
- * (has [sourceUid] and avatar fields, [editable] false), or the uncategorized bucket
- * ([id] null, [collapsible] false).
- */
-private data class HomeCategoryGroup(
-    val id: String?,
-    val sourceUid: Int? = null,
-    val sourceAvatarUrl: String? = null,
-    val sourceFallbackAvatarUrl: String? = null,
-    val sourceIsDefault: Boolean = false,
-    val title: String,
-    val items: List<HomeAppItem>,
-    val collapsed: Boolean,
-    val collapsible: Boolean,
-    val editable: Boolean
-)
-
 private data class CategoryNameRequest(
     val category: HomeAppCategory?
-)
-
-private const val SOURCE_CATEGORY_ID_PREFIX = "source_"
-
-/** Visibility flag paired with the tap callback for a single [AlertSnackbar] slot. */
-@Immutable
-data class AlertState(val visible: Boolean, val onShow: () -> Unit)
-
-/** Transient state driving the bundle-update progress snackbar. */
-@Immutable
-data class BundleUpdateState(
-    val visible: Boolean,
-    val status: BundleUpdateStatus,
-    val progress: PatchBundleRepository.BundleUpdateProgress?
-)
-
-/** Aggregate of all notification-strip inputs, grouped by alert. */
-@Immutable
-data class HomeNotificationsUi(
-    val managerUpdate: AlertState,
-    val blockedSources: AlertState,
-    val metadataErrors: AlertState,
-    val meteredSkipped: AlertState,
-    val bundleUpdate: BundleUpdateState
 )
 
 /** Visible and hidden app lists with their loading state. */
@@ -530,411 +480,6 @@ private fun AppGroupingFooter(
             onModeChange = onModeChange,
             modifier = modifier
         )
-    }
-}
-
-/**
- * Section 1: Unified notifications overlay component.
- * Handles both manager update and bundle update notifications.
- */
-@Composable
-fun NotificationsOverlay(
-    notifications: HomeNotificationsUi,
-    modifier: Modifier = Modifier
-) {
-    Box(
-        modifier = modifier,
-        contentAlignment = Alignment.TopCenter
-    ) {
-        Column(
-            modifier = Modifier.fillMaxWidth(),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            // Blocked source alert takes priority and cannot be dismissed while the block persists
-            AlertSnackbar(
-                visible = notifications.blockedSources.visible,
-                level = AlertLevel.Error,
-                icon = Icons.Outlined.Block,
-                title = stringResource(R.string.home_blocked_source_title),
-                subtitle = stringResource(R.string.home_blocked_source_subtitle),
-                onShowDetails = notifications.blockedSources.onShow,
-                swipeEnabled = false,
-                modifier = Modifier.fillMaxWidth()
-            )
-
-            AlertSnackbar(
-                visible = notifications.meteredSkipped.visible,
-                level = AlertLevel.Warning,
-                icon = Icons.Outlined.SignalCellularAlt,
-                title = stringResource(R.string.home_metered_skipped_title),
-                subtitle = stringResource(R.string.home_metered_skipped_subtitle),
-                onShowDetails = notifications.meteredSkipped.onShow,
-                modifier = Modifier.fillMaxWidth()
-            )
-
-            AlertSnackbar(
-                visible = notifications.metadataErrors.visible,
-                level = AlertLevel.Warning,
-                icon = Icons.Outlined.CloudOff,
-                title = stringResource(R.string.home_metadata_errors_title),
-                subtitle = stringResource(R.string.home_metadata_errors_subtitle),
-                onShowDetails = notifications.metadataErrors.onShow,
-                modifier = Modifier.fillMaxWidth()
-            )
-
-            AlertSnackbar(
-                visible = notifications.managerUpdate.visible,
-                level = AlertLevel.Info,
-                icon = Icons.Outlined.Update,
-                title = stringResource(R.string.home_update_available),
-                subtitle = stringResource(R.string.home_update_available_subtitle),
-                onShowDetails = notifications.managerUpdate.onShow,
-                modifier = Modifier.fillMaxWidth()
-            )
-
-            BundleUpdateSnackbar(
-                visible = notifications.bundleUpdate.visible,
-                status = notifications.bundleUpdate.status,
-                progress = notifications.bundleUpdate.progress,
-                modifier = Modifier.fillMaxWidth()
-            )
-        }
-    }
-}
-
-/** Semantic level of an [AlertSnackbar], resolved into a Material color pair at call time. */
-enum class AlertLevel { Info, Warning, Error, Success }
-
-private data class AlertColorPair(val container: Color, val content: Color)
-
-@Composable
-private fun alertColorsFor(level: AlertLevel): AlertColorPair = when (level) {
-    AlertLevel.Info -> AlertColorPair(
-        MaterialTheme.colorScheme.tertiaryContainer,
-        MaterialTheme.colorScheme.onTertiaryContainer
-    )
-    AlertLevel.Warning -> AlertColorPair(
-        MaterialTheme.colorScheme.secondaryContainer,
-        MaterialTheme.colorScheme.onSecondaryContainer
-    )
-    AlertLevel.Error -> AlertColorPair(
-        MaterialTheme.colorScheme.errorContainer,
-        MaterialTheme.colorScheme.onErrorContainer
-    )
-    AlertLevel.Success -> AlertColorPair(
-        MaterialTheme.colorScheme.primaryContainer,
-        MaterialTheme.colorScheme.onPrimaryContainer
-    )
-}
-
-/**
- * Dismissible card-style alert used for the home-screen notification strip. Swipe-dismiss clears
- * the alert for the current session; it reappears next launch while [visible] stays true.
- */
-@Composable
-fun AlertSnackbar(
-    visible: Boolean,
-    level: AlertLevel,
-    icon: ImageVector,
-    title: String,
-    subtitle: String,
-    onShowDetails: () -> Unit,
-    modifier: Modifier = Modifier,
-    swipeEnabled: Boolean = true
-) {
-    val colors = alertColorsFor(level)
-    val dismissed = remember { mutableStateOf(false) }
-    LaunchedEffect(visible) { if (visible) dismissed.value = false }
-
-    val swipeState = rememberSwipeToDismissBoxState(
-        positionalThreshold = { totalDistance -> totalDistance * 0.4f }
-    )
-    LaunchedEffect(swipeState.currentValue) {
-        if (swipeEnabled && swipeState.currentValue != SwipeToDismissBoxValue.Settled) {
-            dismissed.value = true
-        }
-    }
-
-    AnimatedVisibility(
-        visible = visible && !dismissed.value,
-        enter = MorpheAnimations.slideUpFadeEnter,
-        exit = MorpheAnimations.slideUpFadeExit,
-        modifier = modifier
-    ) {
-        SwipeToDismissBox(
-            state = swipeState,
-            backgroundContent = {},
-            enableDismissFromStartToEnd = swipeEnabled,
-            enableDismissFromEndToStart = swipeEnabled
-        ) {
-            Card(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 16.dp),
-                onClick = onShowDetails,
-                colors = CardDefaults.cardColors(containerColor = colors.container),
-                elevation = CardDefaults.cardElevation(defaultElevation = 8.dp),
-                shape = RoundedCornerShape(16.dp)
-            ) {
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(16.dp),
-                    horizontalArrangement = Arrangement.spacedBy(12.dp),
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    MorpheIcon(icon = icon, tint = colors.content)
-
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            text = title,
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.Bold,
-                            color = colors.content
-                        )
-                        Text(
-                            text = subtitle,
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = colors.content.copy(alpha = 0.8f)
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Bundle update snackbar.
- */
-@Composable
-fun BundleUpdateSnackbar(
-    visible: Boolean,
-    status: BundleUpdateStatus,
-    progress: PatchBundleRepository.BundleUpdateProgress?,
-    modifier: Modifier = Modifier
-) {
-    val dismissed = remember { mutableStateOf(false) }
-    // Reset when a new update cycle starts
-    LaunchedEffect(visible, status) {
-        if (visible && status == BundleUpdateStatus.Updating) dismissed.value = false
-    }
-
-    // Allow swipe only for terminal states - don't let user dismiss an in-progress update
-    val swipeable = status != BundleUpdateStatus.Updating
-
-    val swipeState = rememberSwipeToDismissBoxState(
-        positionalThreshold = { totalDistance -> totalDistance * 0.4f }
-    )
-    LaunchedEffect(swipeState.currentValue) {
-        if (swipeable && swipeState.currentValue != SwipeToDismissBoxValue.Settled) {
-            dismissed.value = true
-        }
-    }
-
-    AnimatedVisibility(
-        visible = visible && !dismissed.value,
-        enter = MorpheAnimations.slideUpFadeEnter,
-        exit = MorpheAnimations.slideUpFadeExit,
-        modifier = modifier
-    ) {
-        SwipeToDismissBox(
-            state = swipeState,
-            backgroundContent = {},
-            enableDismissFromStartToEnd = swipeable,
-            enableDismissFromEndToStart = swipeable
-        ) {
-            BundleUpdateSnackbarContent(status = status, progress = progress)
-        }
-    }
-}
-
-/**
- * Snackbar content with status indicator.
- */
-@Composable
-private fun BundleUpdateSnackbarContent(
-    status: BundleUpdateStatus,
-    progress: PatchBundleRepository.BundleUpdateProgress?
-) {
-    val fraction = if (progress == null || progress.total == 0) 0f
-                   else progress.completed.toFloat() / progress.total
-
-    val downloadFraction = progress?.bytesTotal
-        ?.takeIf { it > 0L }
-        ?.let { progress.bytesRead.toFloat() / it }
-        ?: 0f
-
-    val isDownloading = progress?.phase == PatchBundleRepository.BundleUpdatePhase.Downloading &&
-            downloadFraction > 0f
-    val displayProgress = if (isDownloading) downloadFraction else fraction
-
-    val containerColor by animateColorAsState(
-        targetValue = when (status) {
-            BundleUpdateStatus.Success -> MaterialTheme.colorScheme.primaryContainer
-            BundleUpdateStatus.Warning -> MaterialTheme.colorScheme.secondaryContainer
-            BundleUpdateStatus.Error -> MaterialTheme.colorScheme.errorContainer
-            BundleUpdateStatus.Updating -> MaterialTheme.colorScheme.surfaceVariant
-        },
-        animationSpec = tween(durationMillis = 300),
-        label = "containerColor"
-    )
-
-    val contentColor by animateColorAsState(
-        targetValue = when (status) {
-            BundleUpdateStatus.Success -> MaterialTheme.colorScheme.onPrimaryContainer
-            BundleUpdateStatus.Warning -> MaterialTheme.colorScheme.onSecondaryContainer
-            BundleUpdateStatus.Error -> MaterialTheme.colorScheme.onErrorContainer
-            BundleUpdateStatus.Updating -> MaterialTheme.colorScheme.onSurfaceVariant
-        },
-        animationSpec = tween(durationMillis = 300),
-        label = "contentColor"
-    )
-
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 16.dp),
-        colors = CardDefaults.cardColors(containerColor = containerColor),
-        elevation = CardDefaults.cardElevation(defaultElevation = 8.dp),
-        shape = RoundedCornerShape(16.dp)
-    ) {
-        Column {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(16.dp),
-                horizontalArrangement = Arrangement.spacedBy(12.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                // Icon based on status
-                Crossfade(targetState = status, label = "snackbarIcon") { s ->
-                    when (s) {
-                        BundleUpdateStatus.Success -> MorpheIcon(
-                            icon = Icons.Outlined.CheckCircle,
-                            tint = contentColor
-                        )
-                        BundleUpdateStatus.Warning -> MorpheIcon(
-                            icon = Icons.Outlined.SignalCellularAlt,
-                            tint = contentColor
-                        )
-                        BundleUpdateStatus.Error -> MorpheIcon(
-                            icon = Icons.Outlined.Warning,
-                            tint = contentColor
-                        )
-                        BundleUpdateStatus.Updating -> CircularProgressIndicator(
-                            modifier = Modifier.size(24.dp),
-                            strokeWidth = 2.5.dp,
-                            color = contentColor
-                        )
-                    }
-                }
-
-                // Text content
-                Column(modifier = Modifier.weight(1f)) {
-                    Crossfade(targetState = status, label = "snackbarTitle") { s ->
-                        Text(
-                            text = when (s) {
-                                BundleUpdateStatus.Success -> stringResource(R.string.home_update_success)
-                                BundleUpdateStatus.Warning -> stringResource(R.string.home_update_skipped_metered)
-                                BundleUpdateStatus.Error -> stringResource(R.string.home_update_error)
-                                BundleUpdateStatus.Updating -> stringResource(R.string.home_updating_sources)
-                            },
-                            style = MaterialTheme.typography.titleMedium,
-                            fontWeight = FontWeight.Bold,
-                            color = contentColor
-                        )
-                    }
-
-                    val subtitleColor = contentColor.copy(alpha = 0.8f)
-
-                    if (status == BundleUpdateStatus.Warning) {
-                        Text(
-                            text = stringResource(R.string.home_update_skipped_metered_subtitle),
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = subtitleColor
-                        )
-                    }
-
-                    if (status == BundleUpdateStatus.Updating && progress != null) {
-                        val totalMb = (progress.bytesTotal ?: 0L).toFloat() / (1024 * 1024)
-                        val readMb = progress.bytesRead.toFloat() / (1024 * 1024)
-                        val percent = (downloadFraction * 100).toInt()
-                        val (subtitleKey, subtitle) = when {
-                            progress.total > 1 && isDownloading -> 1 to stringResource(
-                                R.string.home_update_bundle_count_with_bytes,
-                                progress.completed, progress.total, readMb, totalMb, percent
-                            )
-                            progress.total > 1 -> 2 to stringResource(
-                                R.string.home_update_bundle_count,
-                                progress.completed, progress.total
-                            )
-                            isDownloading -> 3 to stringResource(
-                                R.string.home_update_download_progress,
-                                readMb, totalMb, percent.toString()
-                            )
-                            progress.currentBundleName != null -> 4 to progress.currentBundleName
-                            else -> 0 to null
-                        }
-                        AnimatedContent(
-                            targetState = subtitleKey to subtitle,
-                            contentKey = { it.first },
-                            transitionSpec = MorpheAnimations.fadeCrossfade(200),
-                            label = "subtitle"
-                        ) { (_, text) ->
-                            if (text != null) {
-                                Text(
-                                    text = text,
-                                    style = MaterialTheme.typography.bodyMedium,
-                                    color = subtitleColor,
-                                    maxLines = 1,
-                                    overflow = TextOverflow.Ellipsis
-                                )
-                            }
-                        }
-                        AnimatedVisibility(visible = progress.activeNames.isNotEmpty()) {
-                            Crossfade(
-                                targetState = progress.activeNames.joinToString(", "),
-                                label = "activeNames"
-                            ) { names ->
-                                Text(
-                                    text = names,
-                                    style = MaterialTheme.typography.bodySmall,
-                                    color = contentColor.copy(alpha = 0.6f),
-                                    maxLines = 1,
-                                    overflow = TextOverflow.Ellipsis
-                                )
-                            }
-                        }
-                    }
-
-                    if (status == BundleUpdateStatus.Success) {
-                        Text(
-                            text = stringResource(R.string.home_update_success_subtitle),
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = subtitleColor
-                        )
-                    }
-                    if (status == BundleUpdateStatus.Error) {
-                        Text(
-                            text = stringResource(R.string.home_update_error_subtitle),
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = subtitleColor
-                        )
-                    }
-                }
-            }
-
-            // Progress bar for updating state
-            if (status == BundleUpdateStatus.Updating && displayProgress > 0f) {
-                LinearProgressIndicator(
-                    progress = { displayProgress },
-                    modifier = Modifier.fillMaxWidth(),
-                    color = MaterialTheme.colorScheme.primary,
-                    trackColor = MaterialTheme.colorScheme.surfaceVariant
-                )
-            }
-        }
     }
 }
 
@@ -1704,320 +1249,6 @@ private fun ShowHiddenAppsButton(
     )
 }
 
-/**
- * Bucket [items] into the user's custom categories plus an uncategorized tail.
- * [ignoreCollapsed] is used by the search flow to force every group open so matches are
- * visible; when true, empty categories are also dropped from the result.
- */
-private fun buildHomeCategoryGroups(
-    items: List<HomeAppItem>,
-    categoryState: HomeAppCategoryState,
-    uncategorizedTitle: String,
-    ignoreCollapsed: Boolean
-): List<HomeCategoryGroup> {
-    val assigned = items.groupBy { item -> categoryState.assignments[item.packageName] }
-    val groups = categoryState.categories.mapNotNull { category ->
-        val categoryItems = assigned[category.id].orEmpty()
-        if (categoryItems.isEmpty() && ignoreCollapsed) return@mapNotNull null
-        HomeCategoryGroup(
-            id = category.id,
-            title = category.name,
-            items = categoryItems,
-            collapsed = !ignoreCollapsed && category.collapsed,
-            collapsible = true,
-            editable = true
-        )
-    }
-
-    val knownCategoryIds = categoryState.categories.mapTo(mutableSetOf()) { it.id }
-    val uncategorizedItems = items.filter { item ->
-        categoryState.assignments[item.packageName] !in knownCategoryIds
-    }
-    val uncategorizedGroup = uncategorizedItems.takeIf { it.isNotEmpty() }?.let {
-        HomeCategoryGroup(
-            id = null,
-            title = uncategorizedTitle,
-            items = it,
-            collapsed = false,
-            collapsible = false,
-            editable = false
-        )
-    }
-
-    return groups + listOfNotNull(uncategorizedGroup)
-}
-
-/**
- * Bucket [items] into their owning source groups plus an uncategorized tail for anything
- * a source did not claim. Sources are consumed in the order given, so if the same package
- * is declared by multiple sources it lands in the first one only. [ignoreCollapsed]
- * behaves as in [buildHomeCategoryGroups].
- */
-private fun buildHomeSourceGroups(
-    items: List<HomeAppItem>,
-    sourceGroups: List<HomeAppSourceGroup>,
-    uncategorizedTitle: String,
-    ignoreCollapsed: Boolean
-): List<HomeCategoryGroup> {
-    val usedPackages = mutableSetOf<String>()
-
-    val groups = sourceGroups.mapNotNull { sourceGroup ->
-        val sourceItems = items
-            .filter { item ->
-                item.packageName in sourceGroup.packageNames &&
-                        usedPackages.add(item.packageName)
-            }
-
-        if (sourceItems.isEmpty()) {
-            null
-        } else {
-            HomeCategoryGroup(
-                id = "$SOURCE_CATEGORY_ID_PREFIX${sourceGroup.uid}",
-                sourceUid = sourceGroup.uid,
-                sourceAvatarUrl = sourceGroup.avatarUrl,
-                sourceFallbackAvatarUrl = sourceGroup.fallbackAvatarUrl,
-                sourceIsDefault = sourceGroup.isDefault,
-                title = sourceGroup.name,
-                items = sourceItems,
-                collapsed = !ignoreCollapsed && sourceGroup.collapsed,
-                collapsible = sourceGroup.collapsible,
-                editable = false
-            )
-        }
-    }
-
-    val uncategorizedItems = items.filter { item -> item.packageName !in usedPackages }
-    val uncategorizedGroup = uncategorizedItems.takeIf { it.isNotEmpty() }?.let {
-        HomeCategoryGroup(
-            id = null,
-            title = uncategorizedTitle,
-            items = it,
-            collapsed = false,
-            collapsible = false,
-            editable = false
-        )
-    }
-
-    return groups + listOfNotNull(uncategorizedGroup)
-}
-
-/**
- * Row that titles one collapsible section in the home list. Renders the folder or source
- * icon, group title, item count, an optional overflow menu for editable groups (rename or
- * delete), an optional drag-handle slot supplied by the caller (see
- * [CategoryHeaderDragHandle]), and a chevron mirroring the collapse state. Non-collapsible
- * groups (uncategorized, default source) omit the chevron and are not clickable as a whole.
- */
-@Composable
-private fun HomeCategoryHeader(
-    group: HomeCategoryGroup,
-    onToggle: () -> Unit,
-    onRename: () -> Unit,
-    onDelete: () -> Unit,
-    modifier: Modifier = Modifier,
-    dragHandle: (@Composable () -> Unit)? = null
-) {
-    var menuExpanded by remember { mutableStateOf(false) }
-    val countText = pluralStringResource(
-        R.plurals.home_category_app_count,
-        group.items.size,
-        group.items.size.toString()
-    )
-    val headerShape = RoundedCornerShape(8.dp)
-    val containerColor = MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.16f)
-    val contentColor = MaterialTheme.colorScheme.onSurface
-    val mutedContentColor = MaterialTheme.colorScheme.onSurfaceVariant
-    val isSourceGroup = group.sourceUid != null
-    val leadingIcon = when {
-        group.collapsed -> Icons.Outlined.Folder
-        else -> Icons.Outlined.FolderOpen
-    }
-
-    Surface(
-        modifier = modifier
-            .fillMaxWidth()
-            .then(if (group.collapsible) Modifier.clickable(onClick = onToggle) else Modifier),
-        color = containerColor,
-        contentColor = contentColor,
-        shape = headerShape,
-        border = BorderStroke(1.dp, MaterialTheme.colorScheme.outline.copy(alpha = 0.10f))
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .heightIn(min = 44.dp)
-                .padding(start = 10.dp, end = 6.dp, top = 6.dp, bottom = 6.dp),
-            verticalAlignment = Alignment.CenterVertically,
-            horizontalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            if (isSourceGroup) {
-                SourceCategoryIcon(
-                    group = group,
-                    modifier = Modifier.size(22.dp)
-                )
-            } else {
-                Icon(
-                    imageVector = leadingIcon,
-                    contentDescription = null,
-                    modifier = Modifier.size(18.dp),
-                    tint = mutedContentColor
-                )
-            }
-
-            Row(
-                modifier = Modifier.weight(1f),
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(6.dp)
-            ) {
-                Text(
-                    text = group.title,
-                    style = MaterialTheme.typography.titleSmall,
-                    fontWeight = FontWeight.SemiBold,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    modifier = Modifier.weight(1f, fill = false)
-                )
-                Text(
-                    text = countText,
-                    style = MaterialTheme.typography.labelSmall,
-                    color = mutedContentColor,
-                    maxLines = 1
-                )
-            }
-
-            if (group.editable) {
-                Box {
-                    IconButton(
-                        onClick = { menuExpanded = true },
-                        modifier = Modifier.size(32.dp)
-                    ) {
-                        Icon(
-                            imageVector = Icons.Outlined.MoreVert,
-                            contentDescription = stringResource(R.string.more_options),
-                            modifier = Modifier.size(18.dp),
-                            tint = mutedContentColor
-                        )
-                    }
-                    DropdownMenu(
-                        expanded = menuExpanded,
-                        onDismissRequest = { menuExpanded = false }
-                    ) {
-                        DropdownMenuItem(
-                            text = { Text(stringResource(R.string.rename)) },
-                            leadingIcon = { Icon(Icons.Outlined.Edit, contentDescription = null) },
-                            onClick = {
-                                menuExpanded = false
-                                onRename()
-                            }
-                        )
-                        DropdownMenuItem(
-                            text = { Text(stringResource(R.string.delete)) },
-                            leadingIcon = { Icon(Icons.Outlined.Delete, contentDescription = null) },
-                            onClick = {
-                                menuExpanded = false
-                                onDelete()
-                            }
-                        )
-                    }
-                }
-            }
-
-            dragHandle?.invoke()
-
-            if (group.collapsible) {
-                Icon(
-                    imageVector = if (group.collapsed) Icons.Outlined.ExpandMore else Icons.Outlined.ExpandLess,
-                    contentDescription = if (group.collapsed)
-                        stringResource(R.string.expand)
-                    else
-                        stringResource(R.string.collapse),
-                    modifier = Modifier.size(20.dp),
-                    tint = mutedContentColor
-                )
-            }
-        }
-    }
-}
-
-/**
- * Circular avatar for a source-group header. The default (Morphe) source gets the app
- * launcher icon; other sources use the remote avatar from the bundle metadata, with a
- * neutral placeholder when no URL is available.
- */
-@Composable
-private fun SourceCategoryIcon(
-    group: HomeCategoryGroup,
-    modifier: Modifier = Modifier
-) {
-    Surface(
-        modifier = modifier,
-        shape = CircleShape,
-        color = if (group.sourceIsDefault)
-            Color.White
-        else
-            MaterialTheme.colorScheme.surfaceVariant.copy(alpha = 0.55f)
-    ) {
-        when {
-            group.sourceIsDefault -> {
-                val context = LocalContext.current
-                Image(
-                    painter = rememberDrawablePainter(
-                        drawable = AppCompatResources.getDrawable(
-                            context,
-                            R.drawable.ic_launcher_foreground
-                        )
-                    ),
-                    contentDescription = null,
-                    contentScale = ContentScale.Fit,
-                    modifier = Modifier
-                        .fillMaxSize()
-                        .graphicsLayer {
-                            scaleX = 1.5f
-                            scaleY = 1.5f
-                        }
-                )
-            }
-
-            group.sourceAvatarUrl != null -> {
-                RemoteAvatar(
-                    url = group.sourceAvatarUrl,
-                    fallbackUrl = group.sourceFallbackAvatarUrl,
-                    modifier = Modifier.fillMaxSize()
-                )
-            }
-
-            else -> {
-                Icon(
-                    imageVector = Icons.Outlined.Source,
-                    contentDescription = null,
-                    tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.padding(4.dp)
-                )
-            }
-        }
-    }
-}
-
-/**
- * Drag-handle affordance for [HomeCategoryHeader]. Passed as the slot content so the caller
- * can attach a `draggableHandle` from the reorderable library without leaking that state
- * into the header.
- */
-@Composable
-private fun CategoryHeaderDragHandle(modifier: Modifier = Modifier) {
-    Box(
-        modifier = modifier.size(32.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        Icon(
-            imageVector = Icons.Outlined.DragHandle,
-            contentDescription = stringResource(R.string.reorder_list),
-            modifier = Modifier.size(20.dp),
-            tint = MaterialTheme.colorScheme.onSurfaceVariant
-        )
-    }
-}
-
 private fun LazyListScope.hiddenSearchAndShowHiddenItems(
     hiddenAppItems: List<HomeAppItem>,
     filteredHiddenItems: List<HomeAppItem>,
@@ -2379,133 +1610,6 @@ private fun DynamicAppCard(
                     showHideDialog.value = false
                 }
             )
-        }
-    }
-}
-
-/**
- * Animated confirmation bar that slides up from the bottom of the card list
- * when the user is in multi-select mode.
- */
-@Composable
-private fun MultiSelectBar(
-    selectedCount: Int,
-    totalCount: Int,
-    visible: Boolean,
-    onSelectAll: () -> Unit,
-    onDeselectAll: () -> Unit,
-    onAction: () -> Unit,
-    actionIcon: ImageVector,
-    actionContentDescription: String,
-    actionDoneMessage: String,
-    onCancel: () -> Unit,
-    onEnterReorder: () -> Unit,
-    onSaveOrder: () -> Unit,
-    onResetOrder: () -> Unit,
-    onCancelReorder: () -> Unit,
-    modifier: Modifier = Modifier,
-    isReorderMode: Boolean = false,
-    showReorderButton: Boolean = true,
-    actionColors: IconButtonColors = IconButtonDefaults.filledTonalIconButtonColors(
-        containerColor = MaterialTheme.colorScheme.errorContainer,
-        contentColor = MaterialTheme.colorScheme.onErrorContainer
-    ),
-    onMoveToCategory: (() -> Unit)? = null
-) {
-    val effectiveReorderMode = isReorderMode && showReorderButton
-
-    val context = LocalContext.current
-    fun withToast(doneMessage: String, action: () -> Unit): () -> Unit = {
-        context.toast(doneMessage)
-        action()
-    }
-
-    val cancelLabel = stringResource(android.R.string.cancel)
-    val reorderListLabel = stringResource(R.string.reorder_list)
-    val reorderListHint = stringResource(R.string.reorder_list_hint)
-    val reorderDone = stringResource(R.string.reorder_done)
-    val resetOrderLabel = stringResource(R.string.reset_order)
-    val resetOrderDone = stringResource(R.string.reset_order_done)
-    val doneLabel = stringResource(R.string.done)
-    val moveToCategoryLabel = stringResource(R.string.home_category_move_to)
-
-    MultiSelectShell(visible = visible, modifier = modifier) {
-        AnimatedContent(
-            targetState = effectiveReorderMode,
-            transitionSpec = MorpheAnimations.fadeCrossfade(200),
-            label = "multibar_mode"
-        ) { inReorder ->
-            if (inReorder) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 16.dp, vertical = 12.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.spacedBy(8.dp)
-                ) {
-                    Text(
-                        text = reorderListHint,
-                        style = MaterialTheme.typography.labelMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                    ActionPillRow {
-                        ActionPillButton(
-                            onClick = withToast(resetOrderDone, onResetOrder),
-                            icon = Icons.Outlined.Restore,
-                            contentDescription = resetOrderLabel,
-                            tooltip = resetOrderLabel
-                        )
-                        ActionPillButton(
-                            onClick = onCancelReorder,
-                            icon = Icons.Outlined.Close,
-                            contentDescription = cancelLabel,
-                            tooltip = cancelLabel
-                        )
-                        ActionPillButton(
-                            onClick = withToast(reorderDone, onSaveOrder),
-                            icon = Icons.Outlined.Check,
-                            contentDescription = doneLabel,
-                            tooltip = doneLabel
-                        )
-                    }
-                }
-            } else {
-                SelectionActionBar(
-                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 12.dp),
-                    selectedCount = selectedCount,
-                    totalCount = totalCount,
-                    onSelectAll = onSelectAll,
-                    onDeselectAll = onDeselectAll,
-                    onCancel = onCancel
-                ) {
-                    if (onMoveToCategory != null) {
-                        ActionPillButton(
-                            onClick = onMoveToCategory,
-                            icon = Icons.Outlined.FolderOpen,
-                            contentDescription = moveToCategoryLabel,
-                            tooltip = moveToCategoryLabel,
-                            enabled = selectedCount > 0
-                        )
-                    }
-                    ActionPillButton(
-                        onClick = withToast(actionDoneMessage, onAction),
-                        icon = actionIcon,
-                        contentDescription = actionContentDescription,
-                        tooltip = actionContentDescription,
-                        enabled = selectedCount > 0,
-                        colors = actionColors
-                    )
-                    if (showReorderButton) {
-                        ActionPillButton(
-                            onClick = onEnterReorder,
-                            icon = Icons.Outlined.Reorder,
-                            contentDescription = reorderListLabel,
-                            tooltip = reorderListLabel,
-                            enabled = selectedCount > 0
-                        )
-                    }
-                }
-            }
         }
     }
 }
